Move ThemeAction and ThemeApi into kas::draw
