+++ src/traits.rs
@@ -7,7 +7,6 @@
 
 use std::any::Any;
 use std::fmt;
-use std::ops::DerefMut;
 
 use crate::event::{self, Manager};
 use crate::{dir::Direction, layout, WidgetId, WindowId};
@@ -107,61 +106,4 @@ pub trait Window: Widget<Msg = event::VoidMsg> {
     ///
     /// This allows for actions on destruction, but doesn't need to do anything.
     fn handle_closure(&mut self, _mgr: &mut Manager) {}
-}
-
-/// Return value of [`ThemeApi`] functions
-///
-/// This type is used to notify the toolkit of required updates.
-#[derive(Debug, Clone, Copy, Eq, PartialEq, Ord, PartialOrd)]
-pub enum ThemeAction {
-    /// No action needed
-    #[cfg_attr(not(feature = "internal_doc"), doc(hidden))]
-    None,
-    /// All windows require redrawing
-    #[cfg_attr(not(feature = "internal_doc"), doc(hidden))]
-    RedrawAll,
-    /// Theme sizes have changed
-    ///
-    /// This implies that per-window theme data must be updated
-    /// (via [`kas-theme::Theme::update_window`]) and all widgets resized.
-    #[cfg_attr(not(feature = "internal_doc"), doc(hidden))]
-    ThemeResize,
-}
-
-/// Interface through which a theme can be adjusted at run-time
-///
-/// All methods return a [`ThemeAction`] to enable correct action when a theme
-/// is updated via [`Manager::adjust_theme`]. When adjusting a theme before
-/// the UI is started, this return value can be safely ignored.
-pub trait ThemeApi {
-    /// Set font size
-    ///
-    /// Units: Points per Em (standard unit of font size)
-    fn set_font_size(&mut self, pt_size: f32) -> ThemeAction;
-
-    /// Change the colour scheme
-    ///
-    /// If no scheme by this name is found the scheme is left unchanged.
-    // TODO: revise scheme identification and error handling?
-    fn set_colours(&mut self, _scheme: &str) -> ThemeAction;
-
-    /// Switch the theme
-    ///
-    /// Most themes do not react to this method; `kas_theme::MultiTheme` uses
-    /// it to switch themes.
-    fn set_theme(&mut self, _theme: &str) -> ThemeAction {
-        ThemeAction::None
-    }
-}
-
-impl<T: ThemeApi> ThemeApi for Box<T> {
-    fn set_font_size(&mut self, size: f32) -> ThemeAction {
-        self.deref_mut().set_font_size(size)
-    }
-    fn set_colours(&mut self, scheme: &str) -> ThemeAction {
-        self.deref_mut().set_colours(scheme)
-    }
-    fn set_theme(&mut self, theme: &str) -> ThemeAction {
-        self.deref_mut().set_theme(theme)
-    }
 }
